Update conveyors to version 0.5.5
Mainly: fix a serious bug in simple conveyors with the SCATTER option; fix
a major memory leak in mpp_util/SHMEM mode on some platforms; and add a
degenerate (non-parallel) build mode that needs no communication library.

By virtue of testing on MacOS with clang, fix some bugs caused by assuming
that size_t and long are 64 bits in size, and avoid the undefined behavior
of the GCC builtins for counting leading and trailing zero bits when
applied to the value 0.

Some fixes could have been transcribed incorrectly so testing on other
platforms is needed.